Inital Commit
This is probably the 10th time I've tried to start this project. This time I've remembered to use git so that everything is documented and hopefully it doesn't fuck up yet again. All praise papa Gates to bless this project with C#-ey goodness and stop it from going wrong yet again