Added content to ControlVars.cs
Added content to ControlVars.cs. Like the last commit we're going a file at a time so that I can tell when it breaks. This is going to be a long and boring couple of hours doing this a file at a time until it either works or breaks. Fortunatley I hate myself enough that I'm prepared to sit here until it does one or the other. Anyway, all praise papa Gates for the continued function of this project